Cambios en ui, consulta bids, asks, se agregan dependencias para injeccion dependencias

+++ app/src/main/java/com/example/criptobitsoproyectwz/data/model/OrderBook/Asks.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class Asks(
     @SerializedName("book") val book: String,
-    @SerializedName("price") val price: Int,
+    @SerializedName("price") val price: Double,
     @SerializedName("amount") val amount: Double
 )
