+++ app/src/main/java/com/example/criptobitsoproyectwz/data/model/OrderBook/Bids.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class Bids(
     @SerializedName("book") val book: String,
-    @SerializedName("price") val price: Int,
+    @SerializedName("price") val price: Double,
     @SerializedName("amount") val amount: Double
 )
